Fix errors Reply and Post

+++ src/main/java/org/filmt/projetagile/posts/service/impl/PostServiceImpl.java
@@ -4,19 +4,15 @@ import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-
-import org.filmt.projetagile.exception.CategoryNotFoundException;
-import org.filmt.projetagile.exception.GroupNotFoundException;
-import org.filmt.projetagile.exception.UserNotFoundException;
+import org.filmt.projetagile.exception.*;
+import org.filmt.projetagile.groups.service.GroupService;
 import org.filmt.projetagile.posts.dao.PostDAO;
 import org.filmt.projetagile.posts.model.Post;
 import org.filmt.projetagile.posts.service.PostService;
+import org.filmt.projetagile.school.service.SchoolService;
 import org.springframework.context.annotation.Primary;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
 import lombok.AllArgsConstructor;
-import org.springframework.web.server.ResponseStatusException;
 
 @Primary
 @Service
@@ -24,6 +20,9 @@ import org.springframework.web.server.ResponseStatusException;
 public class PostServiceImpl implements PostService {
 
     private final PostDAO postDao;
+
+    GroupService groupService ;
+    SchoolService schoolService ;
 
     @Override
     public Optional<Post> getPostById(final String postId) {
@@ -48,11 +47,19 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public List<Post> getPostsBySchoolIdAndTitle(final String schoolId, final String title) {
+
+        if (schoolService.getSchoolById(schoolId).isEmpty()) {
+            throw SchoolNotFoundException.genericById(schoolId);
+        }
+
         return postDao.getPostsBySchoolIdAndTitle(schoolId, title);
     }
 
     @Override
     public List<Post> getPostsByGroupIdAndTitle(final String groupId, final String title) {
+        if (groupService.getGroupById(groupId).isEmpty()) {
+            throw GroupNotFoundException.genericById(groupId);
+        }
         return postDao.getPostsByGroupIdAndTitle(groupId, title);
     }
 
@@ -74,11 +81,21 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Post update(final Post post) {
+        if (groupService.getGroupById(post.getGroupId()).isEmpty()) {
+            throw GroupNotFoundException.genericById(post.getGroupId());
+        }
+        if (getPostById(post.getId()).isEmpty()) {
+            throw PostNotFoundException.genericById(post.getId());
+        }
         return postDao.update(post);
     }
 
     @Override
     public void delete(final String id) {
-        postDao.delete(id);
+        if(getPostById(id).isPresent()) {
+            postDao.delete(id);
+        } else {
+            throw PostNotFoundException.genericById(id);
+        }
     }
 }
