+++ src/main/java/org/filmt/projetagile/replies/service/impl/ReplyServiceImpl.java
@@ -1,8 +1,10 @@
 package org.filmt.projetagile.replies.service.impl;
 
 import lombok.AllArgsConstructor;
+import org.filmt.projetagile.exception.GroupNotFoundException;
 import org.filmt.projetagile.exception.PostNotFoundException;
 import org.filmt.projetagile.exception.ReplyNotFoundException;
+import org.filmt.projetagile.posts.service.PostService;
 import org.filmt.projetagile.replies.dao.ReplyDAO;
 import org.filmt.projetagile.replies.model.Reply;
 import org.filmt.projetagile.replies.service.ReplyService;
@@ -21,6 +23,8 @@ public class ReplyServiceImpl implements ReplyService {
 
     private final ReplyDAO replyDao ;
 
+    PostService postService ;
+
     @Override
     public Optional<Reply> getReplyById(String replyId) {
         return replyDao.getReplyById(replyId);
@@ -30,19 +34,31 @@ public class ReplyServiceImpl implements ReplyService {
     public Reply create(Reply reply) {
         reply.setId(UUID.randomUUID().toString());
         reply.setCreatedAt(new Timestamp(System.currentTimeMillis()));
-        //if (reply.getPostId().isEmpty()) reply.setReplyId(null);
         replyDao.create(reply);
         return reply ;
     }
 
     @Override
     public Reply update(Reply reply) {
+        if(postService.getPostById(reply.getPostId()).isEmpty()) {
+            throw PostNotFoundException.genericById(reply.getPostId());
+        }
+        if (getReplyById(reply.getReplyId()).isEmpty()) {
+            throw ReplyNotFoundException.genericById(reply.getReplyId());
+        }
+        if (getReplyById(reply.getId()).isEmpty()) {
+            throw ReplyNotFoundException.genericById(reply.getId());
+        }
         return replyDao.update(reply);
     }
 
     @Override
     public void delete(String id) {
-        replyDao.delete(id);
+        if(getReplyById(id).isPresent()) {
+            replyDao.delete(id);
+        } else {
+            throw GroupNotFoundException.genericById(id);
+        }
     }
 
     @Override
